Correções

+++ Programas/EP1/src/threads/OldStateCleaner.java
@@ -15,7 +15,7 @@ public class OldStateCleaner extends Thread { //T4
         while (true) {
             Map<String, PeerInfo> foreignersInfo = DAO.getDAO().getForeignsMap();
             foreignersInfo.forEach((k, v) -> {
-                if (System.currentTimeMillis() - foreignersInfo.get(k).getReceiveMoment() > 1500) {
+                if (System.currentTimeMillis() - foreignersInfo.get(k).getReceiveMoment() > 1500 && !k.equals(DAO.getPeerName())) {
                     System.out.println(String.format("Console peer %s: Thread T4 – eliminando estado do peer %s", DAO.getPeerName(), k));
                     DAO.getDAO().removePeerInfo(k);
                 }
